requirements.txt
arrow==1.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
idna==3.3
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
td-watson==2.0.1
urllib3==1.26.7
